# Optimize await promise