--- a/ProtoPromise_Unity/Assets/Plugins/ProtoPromiseTests/Threading/NonPreservedPromiseConcurrencyTests.cs
+++ b/ProtoPromise_Unity/Assets/Plugins/ProtoPromiseTests/Threading/NonPreservedPromiseConcurrencyTests.cs
@@ -409,9 +409,9 @@ namespace ProtoPromiseTests.Threading
             threadHelper.ExecuteMultiActionParallel(
                 () =>
                 {
-                    Await();
+                    Await().Forget();
 
-                    async void Await()
+                    async Promise Await()
                     {
                         try
                         {
@@ -444,9 +444,9 @@ namespace ProtoPromiseTests.Threading
             threadHelper.ExecuteMultiActionParallel(
                 () =>
                 {
-                    Await();
+                    Await().Forget();
 
-                    async void Await()
+                    async Promise Await()
                     {
                         try
                         {