#26 

Master:

|                  Type |          Method |      N | BaseIsPending |      Mean | Code Size |
|---------------------- |---------------- |------- |-------------- |----------:|----------:|
|          AwaitPending | ProtoPromise_V2 | 100000 |             ? | 202.81 ms |  13,241 B |
|         AwaitResolved | ProtoPromise_V2 | 100000 |             ? |  12.13 ms |   2,845 B |

This PR:

|                  Type |          Method |      N | BaseIsPending |      Mean | Code Size |
|---------------------- |---------------- |------- |-------------- |----------:|----------:|
|          AwaitPending | ProtoPromise_V2 | 100000 |             ? | 176.60 ms |  11,524 B |
|         AwaitResolved | ProtoPromise_V2 | 100000 |             ? |  10.27 ms |   1,906 B |